feat: setup page "Code of conduct"
