+++ ui/src/components/info-page/code-of-conduct-page/code-of-conduct-page.tsx
@@ -1,0 +1,6 @@
+import { InfoPage } from '../info-page'
+import markdown from './code-of-conduct.md'
+
+export const CodeOfConductPage = () => (
+  <InfoPage anchorPrefix="chapter" markdown={markdown} />
+)
