+++ ui/src/components/info-page/code-of-conduct-page/code-of-conduct.md
@@ -1,0 +1,51 @@
+# **Code of conduct**
+
+Hi Friends!
+
+Antenna is an interdisciplinary platform to upload, classify, and analyse in-the-wild images of invertebrates for research and conservation efforts.
+
+Antenna fills the data gap to help understand insects and protect biodiversity by enabling the scale-up of data collection with greater spatial, temporal, and taxonomic coverage than has ever been possible. It is an online platform where entomologists, ecologists, and machine learning (ML) and computer scientists collaborate to build trustworthy datasets by implementing and improving upon ML research designed specifically for real-world applications. Antenna cost-effectively and rapidly classifies a wide range of species in a large number of images, building a rich dataset that grows and updates over time.
+
+Antenna is a highly collaborative community and aspires to be a safe and welcoming space for all. Therefore, we have prepared guidelines for our community in which we outline the expected behavior from everyone, as well as steps for reporting unacceptable behavior.
+
+These guidelines coexist harmoniously with our [Terms of service](https://antenna.insectai.org/terms-of-service) and are not intended to replace them in any way.
+
+1. ## **Scope**
+
+These guidelines do not address behavior beyond Antenna.
+
+2. ## **What we expect from you - _The Do’s_**
+
+- Be empathetic, polite and kind toward other people.
+- Be respectful of people’s projects and contributions.
+- Be respectful of differing opinions, viewpoints, and experiences.
+- Use welcoming and inclusive language.
+- Give and gracefully accept constructive feedback.
+- Accept responsibility and apologise to those affected by our mistakes, and learn from the experience.
+- Focus on what is best not just for us as individuals, but for the overall community.
+
+3. ## **Behaviors that are not welcome - _The Don’ts_**
+
+- The use of sexualized language or imagery, and sexual attention or advances of any kind.
+- Trolling, insulting or making derogatory comments, and personal or political attacks.
+- Discrimination against people based on age, race, gender, income, physical ability, country of origin, educational background, or any other trait that they can't control.
+- Public or private harassment.
+- Disruptive behavior.
+- Publishing others' private information, such as a physical or email address, without their explicit permission.
+  Other conduct which could reasonably be considered inappropriate in a professional setting.
+
+4. ## **Conflict Resolution & Reporting**
+
+Healthy debate is always welcomed & encouraged. However, it is never acceptable to be disrespectful or to engage in behavior that violates Antenna’s community guidelines.
+
+If you notice someone going against our community guidelines, you are encouraged to address the behavior directly with those involved. If you are unable to resolve the matter for any reason, or if the behavior involves threatening or harassing, report it to admin.antenna@mila.quebec.
+
+All complaints will be reviewed and investigated promptly and fairly with respect to the privacy and security of the reporter of any incident.
+
+We will use our discretion in determining when and how to follow up on reported incidents, which may range from not taking action to permanent expulsion from our platform. We will notify the accused of the report and provide them an opportunity to discuss it before any action is taken. In potentially harmful situations, such as ongoing harassment or threats to anyone's safety, we may take action without notice.
+
+Also, keep in mind that moderators may, at their sole discretion & without warning, remove content that we deem harmful to Antenna’s community or data quality, including but not limited to : intentionally false content, content created by sockpuppet accounts, hate speech, pornography, etc.
+
+5. ## **Attribution**
+
+This Code of Conduct is inspired from the [Contributor Covenant](https://www.contributor-covenant.org/), [version 2.1](https://www.contributor-covenant.org/version/2/1/code_of_conduct.html).
