+++ ui/src/components/info-page/info-page.tsx
@@ -1,10 +1,72 @@
-import { TERMS_OF_SERVICE_SLUG } from './terms-of-service-page/constants'
-import { TermsOfServicePage } from './terms-of-service-page/terms-of-service-page'
+import { LoadingSpinner } from 'design-system/components/loading-spinner/loading-spinner'
+import { ReactElement, useEffect, useState } from 'react'
+import ReactMarkdown from 'react-markdown'
+import styles from './info-page.module.scss'
 
-export const InfoPage = ({ slug }: { slug: string }) => {
-  if (slug === TERMS_OF_SERVICE_SLUG) {
-    return <TermsOfServicePage />
+export const InfoPage = ({
+  anchorPrefix,
+  markdown,
+}: {
+  anchorPrefix?: string
+  markdown: URL
+}) => {
+  const [markdownContent, setMarkdownContent] = useState<string>()
+
+  useEffect(() => {
+    const loadContent = async () => {
+      const response = await fetch(markdown)
+      const markdownContent = await response.text()
+      setMarkdownContent(markdownContent)
+    }
+
+    loadContent()
+  })
+
+  if (!markdownContent) {
+    return (
+      <div className={styles.wrapper}>
+        <div className={styles.loadingWrapper}>
+          <LoadingSpinner />
+        </div>
+      </div>
+    )
   }
 
-  return null
+  return (
+    <div className={styles.wrapper}>
+      <div className={styles.container}>
+        <ReactMarkdown
+          className={styles.content}
+          components={{
+            ol: (props) => {
+              if (!anchorPrefix) {
+                return <ol {...props} />
+              }
+
+              const start = props.start ?? 1
+              const children = (props.children as ReactElement[]).filter(
+                (element) => element.type === 'li'
+              )
+
+              if (children.length === 1) {
+                const id = `${anchorPrefix}-${start}`
+
+                return (
+                  <a href={`#${id}`}>
+                    <ol id={id} start={start}>
+                      {props.children}
+                    </ol>
+                  </a>
+                )
+              }
+
+              return <ol start={start}>{props.children}</ol>
+            },
+          }}
+        >
+          {markdownContent}
+        </ReactMarkdown>
+      </div>
+    </div>
+  )
 }
