--- ui/src/components/info-page/terms-of-service-page/constants.ts
@@ -1,1 +1,0 @@
-export const TERMS_OF_SERVICE_SLUG = 'terms-of-service'
